cratedb: initiate deserialization into class level
